Started passing some tests finally!

+++ server/src/main/java/service/UserService.java
@@ -1,6 +1,10 @@
 package service;
 
+import dataaccess.AuthDAO;
 import dataaccess.DataAccessException;
+import dataaccess.UserDAO;
+import model.AuthData;
+import model.UserData;
 
 import java.util.UUID;
 
@@ -8,15 +12,21 @@ public class UserService {
     private final UserDAO userDAO;
     private final AuthDAO authDAO;
 
-    public UserService(UserDAO userDAO, AuthDAO authDAO){
+    public UserService(UserDAO userDAO, AuthDAO authDAO) {
         this.userDAO = userDAO;
         this.authDAO = authDAO;
     }
 
-    public AuthData register(String username, String password, String email) throws DataAccessException{
-        if(username == null || password == null || email == null || username.isEmpty() || password.isEmpty() || email.isEmpty()){
+    public AuthData register(String username, String password, String email) throws DataAccessException {
+        if (username == null || password == null || email == null ||
+                username.isEmpty() || password.isEmpty() || email.isEmpty()) {
             throw new DataAccessException("Error: bad request");
         }
+
+        if (userDAO.getUser(username) != null) {
+            throw new DataAccessException("Error: already taken");
+        }
+
         UserData newUser = new UserData(username, password, email);
         userDAO.createUser(newUser);
 
@@ -27,11 +37,12 @@ public class UserService {
         return authData;
     }
 
-    public AuthData login(String username, String password) throws DataAccessException{
-        if(username == null || password == null || username.isEmpty() || password.isEmpty()){
+    public AuthData login(String username, String password) throws DataAccessException {
+        if (username == null || password == null || username.isEmpty() || password.isEmpty()) {
             throw new DataAccessException("Error: bad request");
         }
-        if(!userDAO.authenticateUser(username, password)){
+
+        if (!userDAO.authenticateUser(username, password)) {
             throw new DataAccessException("Error: unauthorized");
         }
 
@@ -39,10 +50,19 @@ public class UserService {
         AuthData authData = new AuthData(authToken, username);
         authDAO.createAuth(authData);
 
-        return  authData;
+        return authData;
     }
 
-    public void logout(String authToken) throws DataAccessException{
+    public void logout(String authToken) throws DataAccessException {
+        if (authToken == null || authToken.isEmpty()) {
+            throw new DataAccessException("Error: bad request");
+        }
+
+        AuthData authData = authDAO.getAuth(authToken);
+        if (authData == null) {
+            throw new DataAccessException("Error: unauthorized");
+        }
+
         authDAO.deleteAuth(authToken);
     }
 }
